fix: rename / move FrameworkExamples2
