--- components/FrameworkExamples/styles.module.css
@@ -1,3 +1,0 @@
-.example-height {
-  height: 17rem;
-}
